Metropolis-Hastings Step

Perform a Metropolis-Hastings step on one or more states.

If we already have proposed states and all log-density factors we can call mh_step().

If we wish to first sample a proposal and then give the step, we can use mh_sampling_step().

Both metropolis_step() and metropolis_sampling_step() are just wrappers for symmetric transition kernels.

Usage

mh_step(
  x_curr,
  x_prop,
  l_curr,
  l_prop,
  lq_c2p = 0,
  lq_p2c = 0,
  do_checks = TRUE
)

metropolis_step(x_curr, x_prop, l_curr, l_prop, do_checks = TRUE)

mh_sampling_step(
  x_curr,
  l_curr,
  l_target,
  ...,
  sampler,
  sampler_args = NULL,
  lq_sampler = NULL,
  lq_sampler_args = NULL,
  do_checks = TRUE
)

metropolis_sampling_step(
  x_curr,
  l_curr,
  l_target,
  ...,
  sampler,
  sampler_args,
  do_checks = TRUE
)

Arguments

x_curr, x_prop

Two vectors or matrices of current and proposed states. See Dimension for how to specify these to correctly identify the dimension of the state space.

l_curr, l_prop

Vectors of values of log-densities, possibly up to proportionality, of the current and proposed states.

lq_c2p, lq_p2c

(Optional) Vectors of transition log-densities, possibly up to proportionality. The suffix _c2p corresponds to transitions from the current towards the proposed state(s), while the suffix _p2c contains the reversed transitions. If these values are ommited (the default), the Hastings ratio becomes standard Metropolis.

do_checks

If TRUE (the default), runs some preliminary checks for arguments validity. It may be set to FALSE if one is sure that the arguments are correct and doesn't want to incurr in slight overhead.

l_target

For mh_sampling_step(), a function that computes the log-density of the target, possibly up to proportionality. It must accept x_curr as its first argument and its returned value must be a valid corresponding l_prop.

...

Arguments to be passed on to the l_target function.

sampler

For mh_sampling_step(), a sampling function whose first argument is compatible with x_curr and returns appropriate x_prop states.

sampler_args

A list of further arguments to be passed on to the sampler

lq_sampler

(Optional) For mh_sampling_step(), a function that computes the transition log-densities for the sampler(). It first two arguments must be conformable with x_curr and x_prop as it will be used to compute the relevant lq_c2p and lq_p2c factors of the MH step. If ommited, it is assumed that it is symmetric and, hence, it's not needed.

lq_sampler_args

A list of further arguments to be passed on to lq_sampler.

Value

A list containing the results of the Metropolis-Hastings step:

x_next: The next values of the chain. This object has the same structure as the input x_curr. For each proposal, it contains the corresponding x_prop or x_curr values depending on their acceptance or rejection.

l_next: A vector of log-density values corresponding to x_next elements. These may be re-used in the next iteration of a chain to avoid wasting resources in what usually is an expensive computation.

accepted: A vector specifying whether or not each of the proposals was accepted or rejected. May be useful for acceptance rate monitoring.

Dimension

We want to perform MH steps on C states, so all log-density inputs are vectors of length C. The only exception is for the transition log-densities that by default are set both to 0 to signal that we are dealing with symmetrical Metropolis steps. Similarly, the accepted and l_next elements of the output list are always vectors of length C.

States, on the other hand, live in a space of dimension d. There are thus 3 big combinations:

C=1: We want to perform a step on a single d-dimensional state. Without loss of generality, x_curr and x_prop are two vectors of length d.

C>1, d>1: We want to perform a step on each of several multidimensional states. Here, x_curr and x_prop are two C by d matrices, so that each state is a row vector.

C>1, d=1: We want to perform a step on each of several 1-dimensional states. Here, all inputs can be vectors of length C or, alternatively, x_curr and x_prop may be two C by 1 matrices.

The output element x_next is always of the same structure as the input x_curr.

The Metropolis-Hastings Ratio

The usual form of the MH acceptance probability \(\alpha = min{1, MH ratio}\), relies on the ratio $$ MH ratio = \pi(x_1) q(x_0|x_1) / \pi(x_0) q(x_1|x_0) $$ to satisfy detail balance. For numerical reasons, we wish to work on the log scale and the ratio becomes $$ MH log-ratio = l(x_1) + lq(x_0|x_1) - l(x_0) - lq(x_1|x_0) $$ Whenever the transition kernel is symmetrical (i.e. \(q(x_0|x_1)=q(x_1|x_0)\)) we can omit those terms from the calculation and recover the original Metropolis et. al ratio. This is the default assumption of the mh_step() function.

Note on proportionality

We want to also take advantage of the cancellation of normalizing constants so that we only need the log-densities up to a constant of proportionality. Note however that this refers to the same constant. While there are usually no problems when it comes to the target density, care must be taken with general transition kernels as they may have different underlying densities whose normalizing constants don't cancel.