Tempered Mixture log densities generator

Unnormalized and normalized log-tempered densities from a mixture of a user defined log-density.

While mixing densities still yields a normalized density, tempering does not do so and requires renormalization. Hence, the direct result of tempering is the unnormalized ulmix_temp. The normalized version lmix_temp is provided mainly for univariate log-densities, where internal numerical integration is used via stats::integrate.

Component Parameters

Same behavior as lmix: ... expects iterable arguments to be swept across all components while shared_args is a list of shared arguments across components. For a more detailed explanation please see the documentation and examples in lmix.

Usage

ulmix_temp(x, beta = 1, w, ldens, ..., shared_args = NULL)

lmix_temp(x, beta = 1, w, ldens, ..., shared_args = NULL, log_z = NULL)

Arguments

x

A quantile or quantiles if ldens allows for vectorization.

beta

Inverse temperature parameter β > 0.

w

A vector of non-negative mixture weights. To be a valid mixture they must sum to 1.

ldens

A function that returns the log-density of the desired common mixture distribution. For lmix_temp see Details for restrictions on multivariate densities.

...

Other arguments passed on to ldens that vary by mixture component (see Component Parameters).

shared_args

List of other arguments passed on to ldens who are shared by all mixture components (see Component Parameters).

log_z

For lmix_temp, normalizing constant. If NULL (the default) it is estimated via stats::integrate. Hence, while optional for univariate mixtures, it is necessary for proper behavior on multivariate mixtures (see Details).

Value

ulmix_temp returns the unnormalized log-density of the mixture, while lmix_temp estimates the normalized value.

Details

When using lmix_temp for general multivariate mixtures, the user should provide the normalizing constant log_z (in the log scale). This parameter can also be useful for efficiency in one-dimensional situations, as one would avoid estimating it at every call. Note, however, that many tasks in the context of the package don't necessarily require the normalized version; this function is provided for completeness and as a tool for those situations were the normalized versions are indeed sought.

See also